requirements.txt
discord.py>=2.0.0
yt-dlp>=2024.1
asyncio
